requirements.txt
pyOpenSSL
pycryptodome